Fix an error when compiling with g++ on linux
The clang compiler on the max treats the fpos_t type as an int and so did not throw an error when calling the TinyTiffReader_seek_pos function with the fpos_t type.  in g++ the fpos_t type is a struct and so it failed.  The solution was to use theTinyTIFFReader_fsetpos(tiff, &pos) function.